+++ p5-server/tests/testdata/dir with space/file with space.js
@@ -1,0 +1,7 @@
+function setup() {
+  createCanvas(windowWidth, windowHeight);
+}
+
+function draw() {
+  circle(mouseX, mouseY, 20);
+}