+++ p5-server/CHANGELOG.md
@@ -5,6 +5,7 @@
 - New command `p5 screenshot FILE` saves the canvas as an image file.
 - `p5 serve` has `--open <BROWSER>` option
 
+- `p5 serve path/to/sketch.html`
 ## [0.7.1]
 
 - Update README